refactor: change naming and module structure
