+++ src/ess/imaging/tools/resolution.py
